Osobne projekty/UltimateScreenSeats/ultimate-screen-seats-backend/requirements.txt
Django
django-cors-headers
django-ninja
django-ninja-jwt[crypto]
pydantic[email]
pytest
psycopg2

python-decouple